--- app/static/css/common.css
@@ -1,1 +1,0 @@
-:root{--card-color-light: #fcfcfc;--card-color-dark: #faf3f3;--card-color-shadow: #dfdfdf;--text-small: 24pt;--text-weather: 16px;--border-radius: 40px;--shadow-distance: 5px}body{background:var(--card-color-light)}.abs{position:absolute}.svg-shadow{-webkit-filter:drop-shadow(0px 4px 6px rgba(0,0,0,.2));filter:drop-shadow(0px 4px 6px rgba(0,0,0,.2))}.text-small{font-size:var(--text-small)}.text-weather{font-size:var(--text-weather)}.bg-dark{background:var(--card-color-dark)}.bg-light{background:var(--card-color-light)}.mono{font-family:monospace}.sans{font-family:sans-serif}.container{position:relative;margin:20px}.flex{display:flex}.flex-row{flex-direction:row}.flex-column{flex-direction:column}.center{justify-content:center;align-items:center}.neu{border-radius:var(--border-radius);box-shadow:var(--shadow-distance) var(--shadow-distance) 20px var(--card-color-shadow),calc(var(--shadow-distance) * -1) calc(var(--shadow-distance) * -1) 20px #ffffff}
