+++ frontend/.gitignore
@@ -1,3 +1,5 @@
-.snowpack
-build
 node_modules
+.DS_Store
+dist
+dist-ssr
+*.local
