--- frontend/babel.config.json
@@ -1,6 +1,0 @@
-{
-  "presets": ["@babel/preset-env", "@babel/preset-typescript"],
-  "targets": {
-    "ios": "9"
-  }
-}
