+++ frontend/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>오늘내일날씨</title>
+  </head>
+  <body>
+    <div class="flex flex-column center app">
+      <div class="flex flex-row center cards"></div>
+    </div>
+    <script type="module" src="/src/main.ts"></script>
+  </body>
+</html>
