+++ frontend/package.json
@@ -1,22 +1,14 @@
 {
+  "name": "frontend-2",
+  "version": "0.0.0",
   "scripts": {
-    "start": "snowpack dev",
-    "build": "snowpack build",
-    "babel": "babel --extensions .ts src/ -d dist/",
-    "test": "echo \"This template does not include a test runner by default.\" && exit 1",
-    "format": "prettier --write \"src/**/*.{ts,js}\"",
-    "lint": "prettier --check \"src/**/*.{ts,js}\""
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.3",
-    "@babel/core": "^7.14.3",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-typescript": "^7.13.0",
-    "@snowpack/plugin-babel": "^2.1.7",
-    "@snowpack/plugin-typescript": "^1.2.1",
-    "@types/snowpack-env": "^2.3.3",
-    "prettier": "^2.2.1",
-    "snowpack": "^3.3.7",
-    "typescript": "^4.2.4"
+    "@vitejs/plugin-legacy": "^1.4.0",
+    "typescript": "^4.2.3",
+    "vite": "^2.3.4"
   }
 }
