--- frontend/public/index.css
@@ -1,36 +1,0 @@
-:root {
-	--card-width: 400px;
-	--card-height: 600px;
-
-	--card-title-width: calc(var(--card-width) * 0.6);
-	--card-title-height: calc(var(--card-height) * 0.1);
-}
-
-.card-title {
-	top: 20px;
-	left: calc(var(--card-width) / 2);
-	transform: translate(-50%, 0);
-	width: var(--card-title-width);
-	height: var(--card-title-height);
-}
-
-.card-body {
-	width: var(--card-width);
-	height: var(--card-height);
-}
-
-.empty-box {
-	height: 25%;
-}
-
-.weather {
-	height: 50%;
-}
-
-.number-wrapper {
-	height: 25%;
-}
-
-.number {
-	width: 50%;
-}
