--- frontend/public/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta
-      name="description"
-      content="Web site created using create-snowpack-app"
-    />
-    <link rel="stylesheet" type="text/css" href="/index.css" />
-    <link rel="stylesheet" type="text/css" href="/common.css" />
-    <title>오늘내일날씨</title>
-  </head>
-  <body>
-    <div class="flex flex-column center app">
-      <div class="flex flex-row center cards"></div>
-    </div>
-    <script type="module" src="dist/index.js"></script>
-  </body>
-</html>
