--- frontend/snowpack.config.mjs
@@ -1,38 +1,0 @@
-/** @type {import("snowpack").SnowpackUserConfig } */
-export default {
-  mount: {
-    public: { url: '/', static: true },
-    dist: { url: '/dist' },
-  },
-  plugins: [
-    [
-      '@snowpack/plugin-typescript',
-      {
-        /* Yarn PnP workaround: see https://www.npmjs.com/package/@snowpack/plugin-typescript */
-        ...(process.versions.pnp ? { tsc: 'yarn pnpify tsc' } : {}),
-      },
-      '@snowpack/plugin-babel',
-      {
-        input: ['.js, .ts'],
-      },
-    ],
-  ],
-  routes: [
-    /* Enable an SPA Fallback in development: */
-    // {"match": "routes", "src": ".*", "dest": "/index.html"},
-  ],
-  optimize: {
-    minify: true,
-    bundle: true,
-  },
-  packageOptions: {
-    /* ... */
-  },
-  devOptions: {
-    /* ... */
-  },
-  buildOptions: {
-    sourcemap: false,
-    /* ... */
-  },
-};
