+++ frontend/tsconfig.json
@@ -1,29 +1,17 @@
 {
-  "include": ["src", "types"],
   "compilerOptions": {
-    "module": "es2015",
-    "target": "es2015",
-    "moduleResolution": "node",
-    "jsx": "preserve",
-    "baseUrl": "./",
-    /* paths - import rewriting/resolving */
-    "paths": {
-      // If you configured any Snowpack aliases, add them here.
-      // Add this line to get types for streaming imports (packageOptions.source="remote"):
-      //     "*": [".snowpack/types/*"]
-      // More info: https://www.snowpack.dev/guides/streaming-imports
-    },
-    "allowSyntheticDefaultImports": true,
-    "importsNotUsedAsValues": "error",
-    /* more strict checking for errors that per-file transpilers like `esbuild` would crash */
-    "isolatedModules": false,
-    /* noEmit - We only use TypeScript for type checking. */
-    "noEmit": true,
-    /* Additional Options */
+    "target": "ESNext",
+    "module": "ESNext",
+    "lib": ["ESNext", "DOM"],
+    "moduleResolution": "Node",
     "strict": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
+    "sourceMap": true,
     "resolveJsonModule": true,
-    "useDefineForClassFields": true
-  }
+    "esModuleInterop": true,
+    "noEmit": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noImplicitReturns": true
+  },
+  "include": ["./src"]
 }
