+++ frontend/vite.config.js
@@ -1,0 +1,9 @@
+import legacy from "@vitejs/plugin-legacy";
+
+export default {
+  plugins: [
+    legacy({
+      targets: ["ios 9"],
+    }),
+  ],
+};
